Development of foundational layer

+++ FoolsEngine/src/FoolsEngine/Foundation/Memory/Scratchpad.cpp
@@ -6,6 +6,6 @@ namespace fe
 {
 	std::byte* Scratchpad::s_Free;
 	uint64_t Scratchpad::s_RollbackFlags;
-	uint32_t Scratchpad::s_Count;
+	uint64_t Scratchpad::s_Count;
 	std::byte Scratchpad::s_Buffer[Scratchpad::s_BufferSize];
 }
